TST/REF: collect tests by method (#38589)
