+++ pandas/tests/indexes/datetimes/test_datetime.py
@@ -8,8 +8,6 @@ import pandas as pd
 from pandas import DataFrame, DatetimeIndex, Index, Timestamp, date_range, offsets
 import pandas._testing as tm
 
-randn = np.random.randn
-
 
 class TestDatetimeIndex:
     def test_time_loc(self):  # GH8667
@@ -355,10 +353,3 @@ class TestDatetimeIndex:
         result = idx.to_frame()
         expected = DataFrame(idx, index=idx)
         tm.assert_frame_equal(result, expected)
-
-    def test_split_non_utc(self):
-        # GH 14042
-        indices = date_range("2016-01-01 00:00:00+0200", freq="S", periods=10)
-        result = np.split(indices, indices_or_sections=[])[0]
-        expected = indices._with_freq(None)
-        tm.assert_index_equal(result, expected)
